// src/types/enums.ts
export enum MediaType {
  Movie = 'movie',
  Game = 'game',
  Album = 'album',
  TVShow = 'tvshow'
}

export enum GamePlatform {
  PC,
  IOS,
  DC = 'dreamcast',
  GBA = 'gameboy advance',
  DS = 'ds',
  '3DS' = '3ds',
  N64 = 'nintendo 64',
  GC = 'gamecube',
  WII = 'wii',
  WIIU = 'wii u',
  SWITCH = 'switch',
  PS1 = 'playstation 1',
  PS2 = 'playstation 2',
  PS3 = 'playstation 3',
  PS4 = 'playstation 4',
  PS5 = 'playstation 5',
  PSP = 'psp',
  VITA = 'playstation vita',
  XBOX = 'xbox',
  X360 = 'xbox 360',
  XBONE = 'xbox one',
  XSX = 'xbox series x',
  STADIA = 'stadia'
}